requirements.txt
Django==4.0.4
djangorestframework==3.13.1
requests==2.27.1
Pillow==9.1.0